Type annotations via types method.
types is meant to be called right after a function declaration. Here is
an example:

repeatStringNTimes = (str, times) ->
  types "string", "number"

CoffeeScript is a good language, but one of the biggest problems is that
passing in incorrect types can silently fail, or fail in areas far
removed from where the bug took place.

The types method is intended to both annotate function declarations with
type information (for greater readability) and catch most
straightforward type errors ahead of time (for faster debugging). We'll
see how useful it proves.