Start big group refactor.
The idea is that Fathom will control some special properties of groups
in the background.

Motivation:
----------

The Map class is going to create a lot of Tile objects. If the user is
responsible for managing all groups, then the Map class should add the
new Tiles into the groups. But that means that the Map class has to be
responsible for groups, which breaks with our abstraction of having
groups be competely user managed.

The only alternative solution if we want the user to continue to control
all groups would be for the Map class to return a list of all new items
to add into groups. But then where do you return these objects?
initialize()? That doesn't work since initialize can't return anything.
(Duh.) From another method? That doesn't work because it's in my
(currently unwritten) style guide that you should never have to have
the implicit knowledge to call another function after you call the
first.

So we've ruled out all possibilities. We refer to the quote from
Sherlock Holmes [1] and realize that our assumptions are wrong: the
correct answer is that Fathom should be aware of objects.

Implications:
----------

Entities will turn into a MasterGroup class; i.e. it will be aware of
all entities.

Entities will add themselves to MasterGroup by default.

Entities will have "__render" and "__update" groups by default.

The __ will indicate that it is private and only seen by Fathom by
default.

The new MasterGroup class will gain the render, update, etc methods that
were tagging on awkwardly onto Entities. Those methods definitely didn't
belong there, and I'm glad I get to move them to a better place.

A new Group class will emerge. The Group class is little more than a
filter on top of MasterGroup. That is, if I say g = new Group "walls",
I'll just be looking at the "walls" segment of MasterGroup.

[1]: Go look it up.