Give every Entity a reference to the master Entities object.
This is a fairly significant refactor, at least from a theoretical
standpoint. In theory, we lose some of the explicitness of threading the
entities object only into functions where you need it. But in practice
all that ends up happening is that we thread Entities all over the place
and we just end up having more complex code.