@size becomes @width and @height=@width.
Before, what I had done was enforced that the height and width of any
Entity was the value, called size.

Now, the default convention is that the arguments to create a Rect are
(x, y, width, height=width). You can of course customize the height of
the rect by passing that in also.